+++ pkg/aws/services/s3/bucket.go
@@ -39,21 +39,21 @@ func (s *Service) CreateBucket(bucketName string) error {
 	return nil
 }
 
-func (s *Service) CreateTags(bucketName string) error {
+func (s *Service) CreateTags(bucketName string, customerTags map[string]string) error {
 	i := &s3.PutBucketTaggingInput{
 		Bucket: aws.String(bucketName),
 		Tagging: &s3.Tagging{
 			TagSet: []*s3.Tag{
 				{
 					Key:   aws.String(key.S3TagOrganization),
-					Value: aws.String(s.scope.ClusterNamespace()),
+					Value: aws.String(util.RemoveOrg(s.scope.ClusterNamespace())),
 				},
 				{
 					Key:   aws.String(key.S3TagCluster),
 					Value: aws.String(s.scope.ClusterName()),
 				},
 				{
-					Key:   aws.String(fmt.Sprintf(key.S3TagCloudProvider, util.RemoveOrg(s.scope.ClusterName()))),
+					Key:   aws.String(fmt.Sprintf(key.S3TagCloudProvider, s.scope.ClusterName())),
 					Value: aws.String("owned"),
 				},
 				{
@@ -64,11 +64,14 @@ func (s *Service) CreateTags(bucketName string) error {
 		},
 	}
 
+	for k, v := range customerTags {
+		i.Tagging.TagSet = append(i.Tagging.TagSet, &s3.Tag{Key: aws.String(k), Value: aws.String(v)})
+	}
+
 	_, err := s.Client.PutBucketTagging(i)
 	if err != nil {
 		return err
 	}
-	s.scope.Info("Created tags", "bucket", bucketName)
 
 	return nil
 }
@@ -91,7 +94,6 @@ func (s *Service) EncryptBucket(bucketName string) error {
 	if err != nil {
 		return err
 	}
-	s.scope.Info("Encrypted bucket", "bucket", bucketName)
 
 	return nil
 }