+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add giantswarm tags to OIDC S3 bucket.
 - Enable encryption for OIDC S3 bucket.
+- Support Customer tags.
 
 ## [0.2.0] - 2022-03-31