+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-04-13
+
 ### Added
 
 - Add giantswarm tags to OIDC S3 bucket.
@@ -36,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-03-04
 
-[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/giantswarm/irsa-operator/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/irsa-operator/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/giantswarm/irsa-operator/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/giantswarm/irsa-operator/releases/tag/v0.1.0